[TAN-8601] Cover the custom page builder in one Cypress spec
